requirements.txt
google-api-python-client>=2.169.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-dateutil>=2.9.0
python-dotenv>=1.0.1
supabase>=2.15.3
